requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.9.0
aiormq==6.4.2
aiosignal==1.3.1
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
babel==2.16.0
bidict==0.23.1
blis==1.0.1
boto3==1.35.49
botocore==1.35.49
CacheControl==0.12.14
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.5
confluent-kafka==2.6.0
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.8
dask==2022.10.2
dnspython==2.3.0
docopt==0.6.2
es_core_news_md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.8.0/es_core_news_md-3.8.0-py3-none-any.whl#sha256=478b8bb3f3e8eb149192f7d80e7cd64f990b7e9bccbc5329df41a57e86326be2
fbmessenger==6.0.0
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
httptools==0.6.4
humanfriendly==10.0
idna==3.10
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
locket==1.0.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.3
mattermostwrapper==2.2
mdurl==0.1.2
ml_dtypes==0.5.0
msgpack==1.1.0
multidict==5.2.0
murmurhash==1.0.10
networkx==2.6.3
numpy==2.0.2
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==20.9
pamqp==3.2.1
partd==1.4.2
pillow==11.0.0
pluggy==1.5.0
portalocker==2.10.1
preshed==3.0.9
prompt-toolkit==3.0.28
propcache==0.2.0
protobuf==4.23.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.9
pydot==1.4.2
Pygments==2.18.0
PyJWT==2.9.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.2.0
pyreadline3==3.5.4
pyrsistent==0.20.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2022.7.1
pywin32==308
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.6.20
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.3
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.12
s3transfer==0.10.3
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.33.2
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
srsly==2.4.8
structlog==23.3.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminaltables==3.1.10
thinc==8.3.2
threadpoolctl==3.5.0
toolz==1.0.0
tqdm==4.66.5
twilio==8.2.2
typer==0.12.5
typing-utils==0.1.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.0.6
wrapt==1.14.1
wsproto==1.2.0
yarl==1.16.0
